safari-534.53-branch/LayoutTests/sputnik/Conformance/12_Statement/12.5_The_if_Statement/S12.5_A2-expected.txt

S12.5_A2

PASS SputnikError: #1: In the "if" Statement eval as Expression is admitted

TEST COMPLETE
